requirements.txt
pdfplumber==0.10.3
pandas==2.2.1
openai==1.82.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
gunicorn==23.0.0
starlette==0.36.3
pydantic==2.6.1
typing-extensions==4.9.0